Alternative script to generate the s2e images
This script allows for a bit more flexibility than regular image
generation. It allows to specify which distro version and which kernel
and combines them in a new image. It is based on a s2e kprobe kernel
module so any vanilla kernel can be easily run in the image. Once the
image is generated, it is possible to re-use cached files and only
update some elements (e.g., the kernel) in case the image is under
active development. This offer quite a decent speedup.

The script is written to respect most of the operations and directories
used by the default image build, however the naming convention has been
changed to support all combinations between distro/kernel pair without
name clashing